Rename folder section to Navigation
