+++ src/components/account/index.js
@@ -3,7 +3,6 @@ import LogInUser from "./LogInUser";
 import LogOutUser from "./LogOutUser";
 import SignUpUser from "./SignUpUser";
 import NoProfilePresent from "./NoProfilePresent";
-import DeleteAccountButton from "./DeleteAccountButton";
 import AccountIcon from "./AcountIcon";
 import PasswordTextField from "./PasswordTextField";
 import AccountInfo from "./AccountInfo";
@@ -15,7 +14,6 @@ export {
     LogOutUser,
     SignUpUser,
     NoProfilePresent,
-    DeleteAccountButton,
     AccountIcon,
     AccountInfo
 }
